feat: merged source
